requirements.txt
asgiref==3.5.2
cssselect2==0.6.0
cycler==0.11.0
Django==3.2
django-classy-tags==3.0.1
django-cms==3.10.0
django-filer==2.2.1
django-formtools==2.3
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.13.4
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-treebeard==4.4
djangocms-admin-style==3.1.1
djangocms-attributes-field==2.1.0
djangocms-picture==4.0.0
djangocms-text-ckeditor==5.0.1
djangocms-video==3.0.0
easy-thumbnails==2.8.1
fonttools==4.33.3
html5lib==1.1
kiwisolver==1.4.3
lxml==4.9.0
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.0
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.10
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
Unidecode==1.1.2
webencodings==0.5.1
wordcloud==1.8.1
